+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.5.0
+
+* AES-256-CBC for encryption and decryption.
+* Change Cipher -> XOR
+
 ## 1.4.0
 
 * AES-256-CBC for encryption and decryption.